requirements.txt
requests
ipwhois
mac-vendor-lookup
python-whois

